+++ packages/terra-clinical-site/webpack.config.js
@@ -15,15 +15,15 @@ module.exports = {
     'terra-clinical': path.resolve(path.join(__dirname, 'src/Index')),
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.(jsx|js)$/,
         exclude: /node_modules/,
-        loader: 'babel-loader',
+        use: 'babel-loader',
       },
       {
         test: /\.json$/,
-        loader: 'json-loader',
+        use: 'json-loader',
       },
       {
         test: /\.(scss|css)$/,
@@ -46,13 +46,11 @@ module.exports = {
       },
       {
         test: /\.md$/,
-        loader: 'raw-loader',
+        use: 'raw-loader',
       },
       {
         test: /\.(png|svg|jpg|gif)$/,
-        use: [
-          'file-loader',
-        ],
+        use: 'file-loader',
       },
     ],
   },